Add files via upload
While working in DaVinci Resolve (DVR), I add markers in blue colour to the video clips. 
I add markers on the individual clips instead of the timeline. 
Go to Index > Show markers > Blur Color > Three dots > Export Index. Save it as .CSV file. 
To delete all the columns in the CSV by just keeping the time start and description
Run the Powershell script and browse to the CSV file
Click ok
It saves a .TXT file in the same location as the CSV file. 
Copy this TXT file and paste it in your youtube description to get the timelines